Dev/cleanup (#3)
* Remove comment section

Remove commented section that holds example for running an 8-node test. We don't need this because we are prompting the user for this information dynamically.

* Change firewall rule creation to standard format

Move the name of the firewall rule to the top line. Add script variable for project identifier.

* Remove the device name specifier when attaching

The explicit naming of a device name for an attached disk does not seem to work. Remove this line and rely on the pre-defined array of names.

* Add disk and VM deletion functions