Add script to remove all testing artifacts (#11)
* Add script to remove all testing artifacts

* Add clean-project.sh

This script will do a full scrub of a project. Can be used when a testing project is first created or when testing is finished. Note that the vm image is also removed (can be included again with the script create-vm-image.sh

* Add script create-vm-image.sh

Creates a VM image as the foundation for all testing. Required to configure the high-performance network drivers (GVNIC).

* Add script test-configuration.sh

This script contains the configuration for tests. It is used by the new family a creation and deletion scripts.

* Update script removing test artifacts

* Refactoring

* Cleanup, added high-performance options